🚚 rename event_bus to amqp
