+++ marketplace-infrastructure/src/amqp/mod.rs
