+++ marketplace-infrastructure/src/amqp/publisher.rs
