+++ marketplace-infrastructure/src/amqp/subscriber.rs
